+++ FocusYou/Services/Intents/GetStreakIntent.swift
@@ -9,10 +9,13 @@ struct GetStreakIntent: AppIntent {
 
     @MainActor
     func perform() async throws -> some IntentResult & ProvidesDialog {
-        let container = try ModelContainer(
-            for: BlockProfile.self, BlockedSite.self, BlockedApp.self,
-            FocusSession.self, BlockSchedule.self
-        )
+        let dataAccess = try AppIntentDataAccess.makeContainer()
+        if let dialog = dataAccess.unavailableDialog {
+            return .result(dialog: "\(dialog)")
+        }
+        guard let container = dataAccess.container else {
+            return .result(dialog: "\(AppIntentDataAccess.dataStoreUnavailableDialog)")
+        }
         let context = container.mainContext
 
         let sessions = try context.fetch(FetchDescriptor<FocusSession>())